requirements.txt
streamlit
arcadepy
python-dotenv
google-generativeai
requests
